+++ ogcapi-stable/ogcapi-tiles/src/main/java/de/ii/ogcapi/tiles/app/PathParameterTileCol.java
@@ -1,0 +1,97 @@
+/*
+ * Copyright 2022 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.ogcapi.tiles.app;
+
+import com.github.azahnen.dagger.annotations.AutoBind;
+import com.google.common.collect.ImmutableList;
+import de.ii.ogcapi.foundation.domain.ExtensionConfiguration;
+import de.ii.ogcapi.foundation.domain.ExternalDocumentation;
+import de.ii.ogcapi.foundation.domain.OgcApiDataV2;
+import de.ii.ogcapi.foundation.domain.OgcApiPathParameter;
+import de.ii.ogcapi.foundation.domain.SchemaValidator;
+import de.ii.ogcapi.foundation.domain.SpecificationMaturity;
+import de.ii.ogcapi.tiles.domain.TilesConfiguration;
+import io.swagger.v3.oas.models.media.IntegerSchema;
+import io.swagger.v3.oas.models.media.Schema;
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.Optional;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
+/**
+ * @title tileCol
+ * @endpoints Dataset Tile, Collection Tile
+ * @langEn The column of the tile at the zoom level in the tiling scheme.
+ * @langDe Die Spalte der Kachel auf der Zoomstufe im Kachelschema.
+ */
+@Singleton
+@AutoBind
+public class PathParameterTileCol implements OgcApiPathParameter {
+
+  protected final SchemaValidator schemaValidator;
+
+  @Inject
+  PathParameterTileCol(SchemaValidator schemaValidator) {
+    this.schemaValidator = schemaValidator;
+  }
+
+  @Override
+  public String getPattern() {
+    return "\\d+";
+  }
+
+  @Override
+  public List<String> getValues(OgcApiDataV2 apiData) {
+    return ImmutableList.of();
+  }
+
+  private final Schema<?> schema = new IntegerSchema().minimum(BigDecimal.ZERO);
+
+  @Override
+  public Schema<?> getSchema(OgcApiDataV2 apiData) {
+    return schema;
+  }
+
+  @Override
+  public SchemaValidator getSchemaValidator() {
+    return schemaValidator;
+  }
+
+  @Override
+  public String getName() {
+    return "tileCol";
+  }
+
+  @Override
+  public String getDescription() {
+    return "Column index of the tile on the selected zoom level. See http://www.maptiler.org/google-maps-coordinates-tile-bounds-projection/ for more information about Level, Row and Column in the Google Maps tiling scheme (WebMercatorQuad). "
+        + "Example: In the WebMercatorQuad tiling scheme Ireland is fully within the tile with the following values: Level 5, Row 10 and Col 15";
+  }
+
+  @Override
+  public boolean isApplicable(OgcApiDataV2 apiData, String definitionPath) {
+    return isEnabledForApi(apiData)
+        && definitionPath.endsWith("/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}");
+  }
+
+  @Override
+  public Class<? extends ExtensionConfiguration> getBuildingBlockConfigurationType() {
+    return TilesConfiguration.class;
+  }
+
+  @Override
+  public Optional<SpecificationMaturity> getSpecificationMaturity() {
+    return TilesBuildingBlock.MATURITY;
+  }
+
+  @Override
+  public Optional<ExternalDocumentation> getSpecificationRef() {
+    return TilesBuildingBlock.SPEC;
+  }
+}