+++ ogcapi-stable/ogcapi-tiles/src/main/java/de/ii/ogcapi/tiles/domain/WmtsCapabilitiesFormatExtension.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2022 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.ogcapi.tiles.domain;
+
+import com.github.azahnen.dagger.annotations.AutoMultiBind;
+import de.ii.ogcapi.common.domain.GenericFormatExtension;
+import de.ii.ogcapi.foundation.domain.ApiRequestContext;
+import de.ii.ogcapi.foundation.domain.OgcApi;
+import de.ii.ogcapi.foundation.domain.OgcApiDataV2;
+import de.ii.ogcapi.tiles.domain.TilesConfiguration.WmtsScope;
+
+@AutoMultiBind
+public interface WmtsCapabilitiesFormatExtension extends GenericFormatExtension {
+
+  @Override
+  default boolean isEnabledForApi(OgcApiDataV2 apiData) {
+    return apiData
+        .getExtension(TilesConfiguration.class)
+        .filter(TilesConfiguration::isEnabled)
+        .filter(cfg -> cfg.getWmts() != WmtsScope.NONE)
+        .isPresent();
+  }
+
+  Object getEntity(WmtsServiceMetadata capabilities, OgcApi api, ApiRequestContext requestContext);
+}